+++ lizard_map/urls.py
@@ -45,6 +45,12 @@ urlpatterns = patterns(
      {},
      'lizard_map.map_location_load_default'),
 
+    # Download map as image
+    (r'^download/',
+     'lizard_map.views.save_map_as_image',
+     {},
+     'lizard_map.views.save_map_as_image'),
+
     # Collages and snippets
     url(r'^collage/(?P<collage_id>\d+)/$',
         'lizard_map.views.collage',